index product on

+++ Myfrecomm/db/migrate/20160211172500_create_products.rb
@@ -3,9 +3,9 @@ class CreateProducts < ActiveRecord::Migration
     create_table :products do |t|
       t.string :purchaser_name
       t.string :item_description
-      t.float, :item_price
-      t.float :purchase_count
-      t.string, :merchant_address
+      t.decimal :item_price
+      t.decimal :purchase_count
+      t.string :merchant_address
       t.string :merchant_name
 
       t.timestamps null: false
